Rename `ARK` to `IsoBase` in python files.
