+++ isobase/core/image_service.py
