+++ isobase/llm/__init__.py
@@ -1,8 +1,8 @@
 #! python3
 # -*- coding: utf-8 -*-
-from ark.llm.providers import BaseLLMClient, OpenAIChat, AnthropicMessages
-from ark.llm.tools import FunctionTool, ToolSet
-from ark.llm.entities import TokenUsage, LLMResponse, ToolCall
+from isobase.llm.providers import BaseLLMClient, OpenAIChat, AnthropicMessages
+from isobase.llm.tools import FunctionTool, ToolSet
+from isobase.llm.entities import TokenUsage, LLMResponse, ToolCall
 
 __all__ = [
     "BaseLLMClient",
