+++ isobase/llm/entities.py
